feat: add the directory rooter functionality.
The executables now know where to search for the keywords by finding the
root directory of the current project. I took inspiration (and a lot of
the code) from the vim-rooter plugin. However, it doesn't change the
vim directory path like vim-rooter. It just returns the raw output of the
without the other bells and whistles.

For now there is a quick and dirty way of getting the results to the
quickfix list. However, there will need to be a finalize results
function eventually that will allow the user to configure the output a
little further.