feat: Add custom arguments with g:chore_executable_arguments
You can now add completely custom arguments to the executables just by
changing the global variable g:chore_executable_arguments. This will NOT
append to the arugments in s:executables but completely make new ones.
So if the user does not specify the --vimgrep arugment, for instance,
the plugin will completely breakdown.

I also moved the chore#init() invocation to the bottom of the autoload
file. Hopefully this will allow me to test it with a plugin manager such
as vim-plug.