feat: add autoloading command and finalize_search()
The plugin file will now autoload the :ChoreOpen command so that I can
test out the plugin with a plugin manager (vim-plug in my case). When I
did this I also added a way of forcefully loading the whole autoload
file if the user wants to. This shouldn't really be a problem but could
be something someone wants in the future.

I also added the finalize_search() function that computes and picks the
right level of "jump" for the qf list. The levels are as follows:

    0 - Chore will push to the qf list but hide it from the user
    1 - (default) Chore will push to the qf list and open/focus the qf list.
    2 - Chore will push to the qf list and open/focus the qf list and jump to it.

Eventually I want to add the possibility of changing the from the
quickfix list to the location list. Although you're technically not
supposed to use the location list for multi-file errors, some users may
want to if their quickfix list is already in use.