feat: starting to implement async with vim8 jobs
Move the search function to its own file so there can be a fallback
function if '+job' isn't available. In this case it will make a synchronous
search, blocking the ui. Hopefully this will never happen but it is a
precation just in case.

There are still problems extracting the output from the job/channel.
Right now all the output is put into an array where each line is a new
index in the array. Somehow, all the results need to be concated into
one string while still keeping the newline delimters.