feat: finish up the basic async search
You can now search asynchronously but right now vim doesn't know which
function to use. Right now, it is hard coded in to just use the async
search function. Next commit I should have some logic to pick either the
async or the default search.

I also added a default greprg (the executable) and a default grepformat
(%f:%l:%m). This is just a precation more then anything but it is worth
noting.