feat: add a check of async/vim8 features
Add a check to pick either the async or default search function. I also
moved the default search function to autoload/chore/default.vim to go
along with the file structure. It just makes more sense that way.

Also add '--word-regexp' to all of the executables so it has to match
exactly the word we are searching for. This is espeically seen in 'BUG',
where it will find 'BUGS' or 'DEBUG'.