feat: ftplugin for the qf window and delete qf item
Add qf specific settings that will load automatically for that file. It
sets some general settings for now. I might want to add more in the
future but for now this will suffice.

Remap the 'dd' key to delete the current entry under the cursor in the
qf window. There still needs to be some work done on this but for the
most part it is done.